transaction improvements

+++ Spendable/Views/Main/Transactions/TransactionRowView.swift
@@ -37,10 +37,10 @@ struct TransactionRowView: View {
                 
                 HStack {
                     Spacer()
-                    if transaction.amount < 0 {
-                        Text("$" + String(format: "%.2f", abs(transaction.amount))).foregroundColor(.red)
+                    if transaction.amount.doubleValue < 0 {
+                        Text("$" + String(format: "%.2f", abs(transaction.amount.doubleValue))).foregroundColor(.red)
                     } else {
-                        Text("+ $" + String(format: "%.2f", transaction.amount)).foregroundColor(.green)
+                        Text("+ $" + String(format: "%.2f", transaction.amount.doubleValue)).foregroundColor(.green)
                     }
                 }
             }
